Adding the configured test project.
If you're curious on how to setup a test project for your plugin with Doctrine models, this is a pretty good summary.

Once all setup, you execute doctrine:build --all from inside your test project as if it were any other project.

And with any plugin testing, be sure to export a variable called SYMFONY that points to a symfony library:
    export SYMFONY=/home/ryan/symfony/lib