Added cache invalidation at the model layer.
This involves calling the application configuration from the model. Used in the way it is here, it's safe.

While it is possible to specify different cache settings in different applications, or to make each application part of the cache key, since this is caching model-level data (which is shared between all applications equally), there would be no reason for this. As long as this rule isn't broken, the cache invalidation should work fine.